// ui/src/generated/api/model/dtoWorkspaceResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoFileResponse } from "./dtoFileResponse";
import type { DtoBuildResponse } from "./dtoBuildResponse";
import type { DtoPackageMetaResponse } from "./dtoPackageMetaResponse";
import type { DtoStatementResponse } from "./dtoStatementResponse";
import type { DtoTestResponse } from "./dtoTestResponse";

export interface DtoWorkspaceResponse {
  files: DtoFileResponse[];
  issues: string[];
  latestBuild?: DtoBuildResponse;
  meta: DtoPackageMetaResponse;
  statements: DtoStatementResponse[];
  tests: DtoTestResponse[];
}